+++ packages/bundlekit-bundler-parcel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundlekit/bundler-parcel
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [29e8509]
+  - @bundlekit/shared-utils@0.0.11
+
 ## 0.0.10
 
 ### Patch Changes