+++ packages/bundlekit-plugin-angular/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundlekit/plugin-angular
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [29e8509]
+  - @bundlekit/shared-utils@0.0.11
+
 ## 0.0.2
 
 ### Patch Changes